nicer way of defining pin assignments
With these macros, it is possible to define which port and pin
is used for a function in one place. Previously, if a function
was moved to a pin on a different port, it was necessary to
hunt down all the places the pin was referenced and change
the port register. The macros take care of this now.